+++ src/main/java/com/north/producoes/controller/dto/response/ForecastResponseDTO.java
@@ -1,31 +1,32 @@
 package com.north.producoes.controller.dto.response;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
 public record ForecastResponseDTO(
     List<ClientForecastDTO> clients,
     List<MonthSummaryDTO> months,
-    Double monthlyTotal,
-    Double annualTotal,
+    BigDecimal monthlyTotal,
+    BigDecimal annualTotal,
     int year
 ) {
     public record ClientForecastDTO(
         Long clientId,
         String clientName,
         String niche,
-        Double monthlyValue,
+        BigDecimal monthlyValue,
         String status,
         Map<String, ClientMonthDTO> monthData
     ) {}
 
-    public record ClientMonthDTO(Double expected, Double received, Double pending) {}
+    public record ClientMonthDTO(BigDecimal expected, BigDecimal received, BigDecimal pending) {}
 
     public record MonthSummaryDTO(
         String monthKey,
         String monthLabel,
-        Double expected,
-        Double received,
-        Double pending
+        BigDecimal expected,
+        BigDecimal received,
+        BigDecimal pending
     ) {}
 }
